+++ src/index.tsx
@@ -1,27 +1,19 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './utils/normalize.css';
-import './index.css';
-import App from './components/App/App';
-import reportWebVitals from './reportWebVitals';
-import { composeEnhancers } from './components/App/App';
-import { createStore } from 'redux'
-import { Provider } from 'react-redux';
-import { rootReducer } from './services/reducers/reducers';
-import { applyMiddleware } from "redux";
-import thunk from "redux-thunk";
-
-const enhancers = composeEnhancers(applyMiddleware(thunk));
-
-const store = createStore(rootReducer, enhancers)
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./utils/normalize.css";
+import "./index.css";
+import App from "./components/App/App";
+import reportWebVitals from "./reportWebVitals";
+import { store } from "./services/store";
+import { Provider } from "react-redux";
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+  document.getElementById("root") as HTMLElement
 );
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-    <App/>
+      <App />
     </Provider>
   </React.StrictMode>
 );
